// src/models/Bot.ts
export class Bot {
    defaultIntentId: string;
    triggerOnJoinIntentId: string;
    speakSlow: string;
    isCommandsEnabled: boolean;
    isCommandsHelpEnabled: boolean;
    isUtterancesEnabled: boolean;
    isInterceptorEnabled: boolean;
    commandPrefix: string;
    isMainEnabled: boolean;
    isTasksEnabled: boolean;
    isIntentsEnabled: boolean;
}